Add docstring to upcoming.py
